+++ test/Raven.CodeAnalysis.Tests/Completion/CompletionServiceMacroTests.cs
@@ -24,6 +24,7 @@ class MacroHost {
             import System.Collections.Immutable.*
             import Raven.CodeAnalysis.Macros.*
 
+            [LocalMacroPlugin]
             class LocalMacroPlugin : IRavenMacroPlugin {
                 val Name: string => "Local"